backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-text-splitters==0.2.4
# Anthropic (Claude Opus / Sonnet) and Google (Gemini) chat backends.
# Both providers degrade gracefully when their API key is unset --
# their models simply don't appear in the /api/models response.
langchain-anthropic==0.1.23
langchain-google-genai==1.0.10

# OpenAI + compatible httpx
openai==1.47.0
httpx==0.27.2

# FAISS (CPU)
faiss-cpu==1.8.0

# Rate limiting (per-IP throttling on auth + chat endpoints)
slowapi==0.1.9

# Document loaders
pypdf==4.3.1
docx2txt==0.8
# Note: TextLoader is built into langchain-community, no extra package needed

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# MySQL + passwords (user accounts)
# PyMySQL needs `cryptography` for MySQL 8+ default auth (caching_sha2_password)
pymysql==1.1.0
cryptography==42.0.8
bcrypt==4.1.2
tiktoken==0.7.0
numpy==1.26.4

# Payments
stripe==9.11.0

# Google Sign-In (ID token verification)
google-auth==2.36.0

# Microsoft Sign-In (ID token verification via JWKS)
PyJWT[crypto]==2.9.0
